Support bfloat16 activation archives; refresh the acceptance checklist
The extraction forward pass runs in bfloat16, so the fp32 archives carry only
bf16 precision and recasting is bit-exact (verified on 160 sampled files across
all four runs; the conversion refuses any non-lossless write). numpy has no
bfloat16, so load_activations must upcast before crossing into numpy — without
this it raises on every file. The upcast is a no-op for existing fp32 archives.

The checklist claimed several stages were stub-tested only; extraction,
confounds, the causal suite and the SAE have all now run for real. Adds the two
items the real run exposed: the causal dose grid never exceeded 0.087 / 0.011 of
the hidden-state norm, and n_seeds=5 reports std exactly 0.0 because the splits
are manifest-fixed and the probe is deterministic.

Also untracks 7 stale __pycache__/*.pyc files that were committed historically.

Tests: 225 passed.

+++ deception_circuits/paper.py
@@ -180,7 +180,11 @@ def load_activations(
         value = torch.load(path, map_location="cpu", weights_only=True)
         if not isinstance(value, torch.Tensor):
             raise ResearchIntegrityError(f"{path} is not a tensor")
-        arr = value.detach().cpu().numpy()
+        # Activations extracted from a bfloat16 forward pass are stored as
+        # bfloat16 (lossless: the fp32 form carried only bf16 precision).
+        # numpy has no bfloat16, so upcast before crossing into numpy. This is
+        # a no-op for float32 archives written by earlier runs.
+        arr = value.detach().cpu().float().numpy()
         if arr.ndim != 2 or not np.isfinite(arr).all():
             raise ResearchIntegrityError(f"{path} must be a finite [layers, hidden] tensor")
         if shape is None:
